+++ server/utils/updateDatabase.js
@@ -2,7 +2,7 @@ const { processQuery } = require("./db");
 
 
 function updateDatabase() {
-    processQuery(`ALTER TABLE \`bleets\` ADD \`pinned\` VARCHAR(255) NOT NULL AFTER \`file_dir\`; ALTER TABLE \`businesses\` ADD \`business_address\` VARCHAR(255) NOT NULL AFTER \`whitelisted\`; `)
+    processQuery(`ALTER TABLE \`cad_info\` ADD \`company_whitelisted\` VARCHAR(255) NOT NULL AFTER \`whitelisted\`; ALTER TABLE \`bleets\` ADD \`pinned\` VARCHAR(255) NOT NULL AFTER \`file_dir\`; ALTER TABLE \`businesses\` ADD \`business_address\` VARCHAR(255) NOT NULL AFTER \`whitelisted\`; `)
         .then(res => {
             console.log("Updated database for latest version");
         }).catch(err => {